requirements.txt
asgiref==3.5.2
Django==3.2.16
django-request-token==2.0
djangorestframework==3.14.0
Markdown==3.4.1
packaging==21.3
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
pylibmc==1.6.3
pyparsing==3.0.9
pytz==2022.6
sorl-thumbnail==12.9.0
sqlparse==0.4.3
